+++ R/InsertTable.R
@@ -29,11 +29,17 @@ getSqlDataTypes <- function(column) {
     return("FLOAT")
   } else {
     if (is.factor(column)) {
-      maxLength <- max(nchar(levels(column)), na.rm = TRUE)
+      maxLength <-
+        max(suppressWarnings(nchar(
+          stringr::str_conv(string = as.character(column), encoding = "UTF-8")
+        )), na.rm = TRUE)
     } else if (all(is.na(column))) {
       maxLength <- NA
     } else {
-      maxLength <- max(nchar(as.character(column)), na.rm = TRUE)
+      maxLength <-
+        max(suppressWarnings(nchar(
+          stringr::str_conv(string = as.character(column), encoding = "UTF-8")
+        )), na.rm = TRUE)
     }
     if (is.na(maxLength) || maxLength <= 255) {
       return("VARCHAR(255)")
@@ -229,7 +235,7 @@ insertTable.default <- function(connection,
   if (dropTableIfExists) {
     createTable <- TRUE
   }
-  if (tempTable & substr(tableName, 1, 1) != "#" & attr(connection, "dbms") != "redshift") {
+  if (tempTable & substr(tableName, 1, 1) != "#" & dbms(connection) != "redshift") {
     tableName <- paste("#", tableName, sep = "")
   }
   if (!is.null(databaseSchema)) {
@@ -254,21 +260,20 @@ insertTable.default <- function(connection,
     }
   }
   isSqlReservedWord(c(tableName, colnames(data)), warn = TRUE)
-  useBulkLoad <- (bulkLoad && connection@dbms %in% c("hive", "redshift") && createTable) ||
-    (bulkLoad && connection@dbms %in% c("pdw", "postgresql") && !tempTable)
-  useCtasHack <- connection@dbms %in% c("pdw", "redshift", "bigquery", "hive") && createTable && nrow(data) > 0 && !useBulkLoad
-
+  useBulkLoad <- (bulkLoad && dbms(connection) %in% c("hive", "redshift") && createTable) ||
+    (bulkLoad && dbms(connection) %in% c("pdw", "postgresql") && !tempTable)
+  useCtasHack <- dbms(connection) %in% c("pdw", "redshift", "bigquery", "hive") && createTable && nrow(data) > 0 && !useBulkLoad
+  if (dbms(connection) == "bigquery" && useCtasHack && is.null(tempEmulationSchema)) {
+    abort("tempEmulationSchema is required to use insertTable with bigquery when inserting into a new table")
+  }
+  
   sqlDataTypes <- sapply(data, getSqlDataTypes)
   sqlTableDefinition <- paste(.sql.qescape(names(data), TRUE, connection@identifierQuote), sqlDataTypes, collapse = ", ")
   sqlTableName <- .sql.qescape(tableName, TRUE, connection@identifierQuote)
   sqlFieldNames <- paste(.sql.qescape(names(data), TRUE, connection@identifierQuote), collapse = ",")
 
   if (dropTableIfExists) {
-    if (tempTable) {
-      sql <- "IF OBJECT_ID('tempdb..@tableName', 'U') IS NOT NULL DROP TABLE @tableName;"
-    } else {
-      sql <- "IF OBJECT_ID('@tableName', 'U') IS NOT NULL DROP TABLE @tableName;"
-    }
+    sql <- "DROP TABLE IF EXISTS @tableName;"
     renderTranslateExecuteSql(
       connection = connection,
       sql = sql,
@@ -279,7 +284,7 @@ insertTable.default <- function(connection,
     )
   }
 
-  if (createTable && !useCtasHack && !(bulkLoad && connection@dbms == "hive")) {
+  if (createTable && !useCtasHack && !(bulkLoad && dbms(connection) == "hive")) {
     sql <- paste("CREATE TABLE ", sqlTableName, " (", sqlTableDefinition, ");", sep = "")
     renderTranslateExecuteSql(
       connection = connection,
@@ -296,13 +301,13 @@ insertTable.default <- function(connection,
       abort("Bulk load credentials could not be confirmed. Please review them or set 'bulkLoad' to FALSE")
     }
     inform("Attempting to use bulk loading...")
-    if (connection@dbms == "redshift") {
+    if (dbms(connection) == "redshift") {
       bulkLoadRedshift(connection, sqlTableName, data)
-    } else if (connection@dbms == "pdw") {
+    } else if (dbms(connection) == "pdw") {
       bulkLoadPdw(connection, sqlTableName, sqlDataTypes, data)
-    } else if (connection@dbms == "hive") {
+    } else if (dbms(connection) == "hive") {
       bulkLoadHive(connection, sqlTableName, sqlFieldNames, data)
-    } else if (connection@dbms == "postgresql") {
+    } else if (dbms(connection) == "postgresql") {
       bulkLoadPostgres(connection, sqlTableName, sqlFieldNames, sqlDataTypes, data)
     }
   } else if (useCtasHack) {
@@ -324,7 +329,7 @@ insertTable.default <- function(connection,
       ")"
     )
     insertSql <- SqlRender::translate(insertSql,
-      targetDialect = connection@dbms,
+      targetDialect = dbms(connection),
       tempEmulationSchema = tempEmulationSchema
     )
     batchSize <- 10000
@@ -371,10 +376,12 @@ insertTable.default <- function(connection,
           return(NULL)
         }
         lapply(1:ncol(data), setColumn, start = start, end = end)
-        if (attr(connection, "dbms") == "bigquery") {
-          rJava::.jcall(batchedInsert, "V", "executeBigQueryBatch")
+        if (dbms(connection) == "bigquery") {
+          if (!rJava::.jcall(batchedInsert, "Z", "executeBigQueryBatch"))
+            stop("Error uploading data")
         } else {
-          rJava::.jcall(batchedInsert, "V", "executeBatch")
+          if (!rJava::.jcall(batchedInsert, "Z", "executeBatch"))
+            stop("Error uploading data")
         }
       }
       if (progressBar) {
@@ -410,7 +417,7 @@ insertTable.DatabaseConnectorDbiConnection <- function(connection,
 
   tableName <- gsub("^#", "", tableName)
   if (!is.null(databaseSchema)) {
-    if (connection@dbms %in% c("sqlite", "sqlite extended")) {
+    if (dbms(connection) %in% c("sqlite", "sqlite extended")) {
       if (tolower(databaseSchema) != "main") {
         abort("Only the 'main' schema exists on SQLite")
       }
@@ -419,14 +426,15 @@ insertTable.DatabaseConnectorDbiConnection <- function(connection,
     }
   }
 
-  if (connection@dbms == "sqlite") {
+  if (dbms(connection) == "sqlite") {
     # Convert dates and datetime to UNIX timestamp:
     for (i in 1:ncol(data)) {
-      if (inherits(data[, i], "Date")) {
-        data[, i] <- as.numeric(as.POSIXct(as.character(data[, i]), origin = "1970-01-01", tz = "GMT"))
+      column <- data[[i]]
+      if (inherits(column, "Date")) {
+        data[, i] <- as.numeric(as.POSIXct(as.character(column), origin = "1970-01-01", tz = "GMT"))
       }
-      if (inherits(data[, i], "POSIXct")) {
-        data[, i] <- as.numeric(as.POSIXct(data[, i], origin = "1970-01-01", tz = "GMT"))
+      if (inherits(column, "POSIXct")) {
+        data[, i] <- as.numeric(as.POSIXct(column, origin = "1970-01-01", tz = "GMT"))
       }
     }
   }