+++ NEWS.md
@@ -1,3 +1,26 @@
+DatabaseConnector 5.1.0
+=======================
+
+Changes:
+
+1. Supporting upcoming version of `Andromeda` using `arrow` backend.
+
+2. Adding new `dbms()` function, which is the preferred way to establish the dbms of a connection.
+
+3. Adding support for Snowflake.
+
+4. Updating RedShift driver to V2.1.0.9. Fixes error when uploading data.
+
+
+Bugfixes:
+
+1. Fix erroneous dates when inserting dates in a tibble into a SQLite database.
+
+2. Ensuring errors when inserting data turn into R errors.
+
+3. Throwing informative error when using `insertTable()` on BigQuery without specifying the `tempEmulationSchema`.
+
+
 DatabaseConnector 5.0.4
 =======================